Testing changes
I am having difficulty getting advice-content to show what I want it
too, though it seems to work when I change it to food-history-content,
indicating that the problem lies elsewhere. This commit is just so I can
keep the git up to date with the changes I have been making just in case
I need to revert.